requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0
pydantic>=2.0.0
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
scipy>=1.11.0
albumentations>=1.3.0
